create auctions ai app
