+++ auctions/urls.py
@@ -1,9 +1,8 @@
 from django.urls import path
 
-from auctions.views import AuctionListView, AuctionDetailView, ChatFormView
+from auctions.views import AuctionListView, AuctionDetailView
 
 urlpatterns = [
     path('auctions', AuctionListView.as_view(), name="auction.list"),
-    path('auctions/<int:pk>', AuctionDetailView.as_view(), name="auction.detail"),
-    path('chat', ChatFormView.as_view(), name='auction.chat')
+    path('auctions/<int:pk>', AuctionDetailView.as_view(), name="auction.detail")
 ]
