+++ static/templates/base.html
@@ -24,6 +24,7 @@
       </a>
       <ul class='flex justify-end items-center'>
         {% block navlinks %}
+          <li><a href="{% url 'auctions_ai.chat' %}" class='text-sm font-medium leading-6 px-5 hover:text-sky-500'>AI</a></li>
           <li><a href="#" class='text-sm font-medium leading-6 px-5 hover:text-sky-500'>Über uns</a></li> <!-- TODO: add about page -->
           <li><a href="{% url 'immobyte.logout' %}" class='text-sm font-medium leading-6 px-5 hover:text-sky-500'>Logout</a></li>
         {% endblock %}
