+++ tailwind.config.js
@@ -2,7 +2,8 @@
 module.exports = {
   content: [
     './static/templates/**/*.{html,js}',
-    './auctions/templates/**/*.{html,js}'
+    './auctions/templates/**/*.{html,js}',
+    './auctions_ai/templates/**/*.{html,js}'
   ],
   theme: {
     extend: {}
